timerフックを表示ロジック用と部分処理用に分離
